backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
pydantic>=2.5.3
pydantic-settings>=2.1.0
python-dateutil>=2.8.2
python-multipart>=0.0.6
openpyxl>=3.1.2
pandas>=2.2.0
pytest>=7.4.4
pytest-asyncio>=0.23.3
httpx>=0.26.0
